hal/aarch64: misrify
In consequence, changes signatures in global cpu.h, pmap.h, timer.h hal
headers and updates the implementation in the rest of architectures.

JIRA: DO-3

+++ hal/armv7m/cpu.c
@@ -250,7 +250,7 @@ char *hal_cpuInfo(char *info)
 }
 
 
-char *hal_cpuFeatures(char *features, unsigned int len)
+char *hal_cpuFeatures(char *features, size_t len)
 {
 	unsigned int n = 0;
 #ifdef CPU_IMXRT
