+++ hal/sparcv8leon/cpu.c
@@ -201,7 +201,7 @@ char *hal_cpuInfo(char *info)
 }
 
 
-char *hal_cpuFeatures(char *features, unsigned int len)
+char *hal_cpuFeatures(char *features, size_t len)
 {
 	unsigned int n = 0;
 	const char *fpu = hal_cpuGetFpuOption();
